Add files via upload
The script is a keylogger that captures key presses and saves them to a log file or sends them via email, it also has some additional features like include timestamps, process name and ascii art. It uses libraries such as pynput, smtplib, threading, datetime, psutil and pyfiglet to perform its tasks. The script prompts the user to enter their email address and password, and starts capturing key presses, storing them in a log variable. It also starts a timer that calls a function to send the log via email or save it to a file, depending on the user's choice, every two minutes. It's important to note that keylogging is illegal in some countries and is considered a violation of privacy, make sure you have proper authorization and be aware of any laws and regulations that apply to your use case.